+++ neoforge/src/main/java/earth/terrarium/athena/api/client/neoforge/ForgeAthenaUtils.java
@@ -32,10 +32,10 @@ public class ForgeAthenaUtils {
 
     private static Vector3f getStartPos(AthenaQuad quad, Direction direction) {
         return switch (direction) {
-            case NORTH -> new Vector3f((1 - quad.right()) * 16f, quad.top() * 16f, quad.depth() * 16f);
-            case SOUTH -> new Vector3f(quad.left() * 16f, quad.top() * 16f, (1-quad.depth()) * 16f);
-            case WEST -> new Vector3f(quad.depth() * 16f, quad.top() * 16f, quad.left() * 16f);
-            case EAST -> new Vector3f((1 - quad.depth()) * 16f, quad.top() * 16f,  (1 - quad.right()) * 16f);
+            case NORTH -> new Vector3f((1 - quad.left()) * 16f, quad.bottom() * 16f, quad.depth() * 16f);
+            case SOUTH -> new Vector3f(quad.right() * 16f, quad.bottom() * 16f, (1-quad.depth()) * 16f);
+            case WEST -> new Vector3f(quad.depth() * 16f, quad.bottom() * 16f, quad.right() * 16f);
+            case EAST -> new Vector3f((1 - quad.depth()) * 16f, quad.bottom() * 16f,  (1 - quad.left()) * 16f);
             case DOWN -> new Vector3f(quad.left() * 16f, quad.depth() * 16f, quad.top() * 16f);
             case UP -> new Vector3f(quad.left() * 16f, (1 - quad.depth()) * 16f, (1 - quad.bottom()) * 16f);
         };
@@ -43,10 +43,10 @@ public class ForgeAthenaUtils {
 
     private static Vector3f getEndPos(AthenaQuad quad, Direction direction) {
         return switch (direction) {
-            case NORTH -> new Vector3f((1 - quad.left()) * 16f, quad.bottom() * 16f, quad.depth() * 16f);
-            case SOUTH -> new Vector3f(quad.right() * 16f, quad.bottom() * 16f, (1 - quad.depth()) * 16f);
-            case WEST -> new Vector3f(quad.depth() * 16f, quad.bottom() * 16f, quad.right() * 16f);
-            case EAST -> new Vector3f((1 - quad.depth()) * 16f, quad.bottom() * 16f, (1 - quad.left()) * 16f);
+            case NORTH -> new Vector3f((1 - quad.right()) * 16f, quad.top() * 16f, quad.depth() * 16f);
+            case SOUTH -> new Vector3f(quad.left() * 16f, quad.top() * 16f, (1 - quad.depth()) * 16f);
+            case WEST -> new Vector3f(quad.depth() * 16f, quad.top() * 16f, quad.left() * 16f);
+            case EAST -> new Vector3f((1 - quad.depth()) * 16f, quad.top() * 16f, (1 - quad.right()) * 16f);
             case DOWN -> new Vector3f(quad.right() * 16f, quad.depth() * 16f, quad.bottom() * 16f);
             case UP -> new Vector3f(quad.right() * 16f, quad.depth() * 16f, (1 - quad.top()) * 16f);
         };